Added a new section for FaaS
I was surprised to find out you left out the entire section on FaaS, which is what a lot of the movement is toward serverless static sites these days... But then a lot has happened this year and last year this was not the rage.

In keeping with your idea of making this list easy to contribute to, (since I found it through the article on CloudCannon), would you be amenable to me creating a static site out of this list? My idea would be to incorporate a static CMS with it so that people without GH could contribute (but whoever doesn't have GH, why would they?) from the website and the community could upvote, recommend, comment on each of these services? If you would like to hack on it with me, that would be great, or I can just work on it by myself...

Cheers,
Tim